requirements.txt
aiogram==3.7.0
python-dotenv==1.0.0
supabase==2.4.0
APScheduler==3.10.4
aiohttp==3.9.5
python-dateutil==2.9.0
pydantic==2.7.0
